requirements.txt
spacy==2.3.2
pymongo==3.11.0
requests==2.24.0
nltk==3.5
Flask==1.1.2
textblob==0.15.3
python-dotenv==0.14.0